WIP: basic hodlr wrapping
two methods of wrapping:

one following RcppEigen package skeleton examples where there are three steps.

step 1: add the function to RcppExports.R
step 2: add the input/output type definitions to RcppExports.cpp
step 3: add the implementation to the hello world class;
an example add for this approach is rcppeigen_matinv

the 2nd option just uses the first and third steps
an example is getEigenvalues and hodlrwrap

not sure of the difference but the 2nd may copy data
and thus be slower.